Alteração em rotas, controllers, para reparar as rotas em mais de um arquivo, bem como os controllers
