+++ para-o-lar/package.json
@@ -15,6 +15,7 @@
     "express": "^4.18.1"
   },
   "devDependencies": {
+    "eslint": "^8.17.0",
     "nodemon": "^2.0.16"
   }
 }
