+++ para-o-lar/src/app.js
@@ -5,10 +5,12 @@ const app = express();
 app.use(express.json());
 app.use(cors());
 
-const aulasRoutes = require("./routes/aulasRoutes.js");
-const historicoRoutes = require("./routes/historicoRoutes.js");
+const cadastrosPrivadosRoutes = require("./routes/cadastrosPrivadosRoutes.js");
+const cadastrosPublicosRoutes = require("./routes/cadastrosPublicosRoutes.js");
+const historicoRoutes = require("./routes/historicosRoutes.js");
 
-app.use("/classes", aulasRoutes);
-app.use("/records", historicoRoutes);
+app.use("/registers/private", cadastrosPrivadosRoutes);
+app.use("/registers/public", cadastrosPublicosRoutes);
+app.use("/records/private", historicoRoutes);
 
 module.exports = app
